+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Disallow @ts-expect-error and @ts-ignore
+
 ## [1.0.0] - 2022-10-26
 
 ### Added